Update to latest version - version_6
Update to the latest version of the game. 

This version is the cleanest version yet. I actually mapped it all out on a whiteboard and then coded it in just a few hours.

This version is very simple, but I think it is a big step in the right direction.

Uses better techniques and cleaner code than all previous versions.